move InvalidStateStoreException to errors package. Add comment about backoff and retry. Update ReadOnlyKeyValueStore and ReadOnlyWindowStore with throws InvalidStateStoreException
