+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/StreamThread.java
@@ -105,7 +105,7 @@ public class StreamThread extends Thread {
 
     private Map<TopicPartition, List<ConsumerRecord<byte[], byte[]>>> standbyRecords;
     private boolean processStandbyRecords = false;
-    private AtomicBoolean stateStoresAvailable = new AtomicBoolean(false);
+    private AtomicBoolean initialized = new AtomicBoolean(false);
 
     final ConsumerRebalanceListener rebalanceListener = new ConsumerRebalanceListener() {
         @Override
@@ -115,7 +115,7 @@ public class StreamThread extends Thread {
                 addStandbyTasks();
                 lastCleanMs = time.milliseconds(); // start the cleaning cycle
                 streamsMetadataState.onChange(partitionAssignor.getPartitionsByHostState(), partitionAssignor.clusterMetadata());
-                stateStoresAvailable.set(true);
+                initialized.set(true);
             } catch (Throwable t) {
                 rebalanceException = t;
                 throw t;
@@ -125,7 +125,7 @@ public class StreamThread extends Thread {
         @Override
         public void onPartitionsRevoked(Collection<TopicPartition> assignment) {
             try {
-                stateStoresAvailable.set(false);
+                initialized.set(false);
                 commitAll();
                 lastCleanMs = Long.MAX_VALUE; // stop the cleaning cycle until partitions are assigned
             } catch (Throwable t) {
@@ -141,8 +141,8 @@ public class StreamThread extends Thread {
         }
     };
 
-    public boolean allStateStoresAvailable() {
-        return stateStoresAvailable.get();
+    public boolean isInitialized() {
+        return initialized.get();
     }
 
     public StreamThread(TopologyBuilder builder,
